refactor(ai-builder): Centralize agent prompts into dedicated directory
- Extract prompts from subgraphs/agents into src/prompts/agents/
- Add index.ts for centralized exports
- Add brief documentation comments explaining each prompt section
