+++ packages/@n8n/ai-workflow-builder.ee/src/agents/supervisor.agent.ts
@@ -4,53 +4,15 @@ import { HumanMessage } from '@langchain/core/messages';
 import { ChatPromptTemplate } from '@langchain/core/prompts';
 import { z } from 'zod';
 
+import {
+	buildSupervisorPrompt,
+	SUPERVISOR_PROMPT_SUFFIX,
+} from '@/prompts/agents/supervisor.prompt';
+
 import type { CoordinationLogEntry } from '../types/coordination';
 import type { SimpleWorkflow } from '../types/workflow';
 import { buildWorkflowSummary } from '../utils/context-builders';
 import { summarizeCoordinationLog } from '../utils/coordination-log';
-
-/**
- * Supervisor Agent Prompt
- *
- * Handles INITIAL routing based on user intent.
- * After initial routing, deterministic routing takes over based on coordination log.
- */
-const SUPERVISOR_PROMPT = `You are a Supervisor that routes user requests to specialist agents.
-
-AVAILABLE AGENTS:
-- discovery: Find n8n nodes for building/modifying workflows
-- builder: Create nodes and connections (requires discovery first for new node types)
-- configurator: Set parameters on EXISTING nodes (no structural changes)
-- responder: Answer questions, confirm completion (TERMINAL)
-
-ROUTING DECISION TREE:
-
-1. Is user asking a question or chatting? → responder
-   Examples: "what does this do?", "explain the workflow", "thanks"
-
-2. Does the request involve NEW or DIFFERENT node types? → discovery
-   Examples:
-   - "Build a workflow that..." (new workflow)
-   - "Use [ServiceB] instead of [ServiceA]" (replacing node type)
-   - "Add [some integration]" (new integration)
-   - "Switch from [ServiceA] to [ServiceB]" (swapping services)
-
-3. Is the request about connecting/disconnecting existing nodes? → builder
-   Examples: "Connect node A to node B", "Remove the connection to X"
-
-4. Is the request about changing VALUES in existing nodes? → configurator
-   Examples:
-   - "Change the URL to https://..."
-   - "Set the timeout to 30 seconds"
-   - "Update the email subject to..."
-
-KEY DISTINCTION:
-- "Use [ServiceB] instead of [ServiceA]" = REPLACEMENT = discovery (new node type needed)
-- "Change the [ServiceA] API key" = CONFIGURATION = configurator (same node, different value)
-
-OUTPUT:
-- reasoning: One sentence explaining your routing decision
-- next: Agent name`;
 
 const systemPrompt = ChatPromptTemplate.fromMessages([
 	[
@@ -58,9 +20,7 @@ const systemPrompt = ChatPromptTemplate.fromMessages([
 		[
 			{
 				type: 'text',
-				text:
-					SUPERVISOR_PROMPT +
-					'\n\nGiven the conversation above, which agent should act next? Provide your reasoning and selection.',
+				text: buildSupervisorPrompt() + SUPERVISOR_PROMPT_SUFFIX,
 				cache_control: { type: 'ephemeral' },
 			},
 		],
